move common spring controllers to common module
