+++ dd-smoke-tests/springboot/src/main/java/datadog/smoketest/springboot/SpringbootApplication.java
@@ -1,11 +1,12 @@
 package datadog.smoketest.springboot;
 
-import datadog.smoketest.springboot.controller.SimpleIastController;
+import datadog.smoketest.springboot.controllers.SimpleIastController;
 import java.lang.management.ManagementFactory;
 import java.util.Collections;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.web.firewall.HttpFirewall;
 import org.springframework.security.web.firewall.StrictHttpFirewall;
@@ -19,6 +20,7 @@ import org.springframework.web.util.UrlPathHelper;
 public class SpringbootApplication {
 
   @Configuration
+  @ComponentScan(basePackages = {"datadog.smoketest.springboot.controllers"})
   public static class WebConfig extends WebMvcConfigurerAdapter {
 
     @Override
